Added logging, made email optional, updated email content
Removed email message when no updates are available as most people don't
want to be spammed by this. Added logic to only send the email when
there is one to send via the SENDEMAIL variable.
Added the hostname to the subject line
Adjusted wording of the message.
Added logging so that it is easy to monitor that the job has been
running and what the results were.